# src/main/webapp/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "PetStore APIs exposed",
    "version" : "1.0.0",
    "title" : "Petstore APIs",
    "contact" : {
      "name" : "Antonio Goncalves",
      "url" : "https://antoniogoncalves.org/",
      "email" : "antonio.goncalves@gmail.com"
    }
  },
  "host" : "localhost:5000",
  "basePath" : "/rest",
...
  "schemes" : [ "http", "https" ],
  "paths" : {
    "/categories" : {
      "get" : {
        "tags" : [ "Category" ],
        "summary" : "Lists all the categories",
        "description" : "",
        "operationId" : "listAllCategories",
        "produces" : [ "application/xml", "application/json" ],
        "parameters" : [ {
          "name" : "start",
          "in" : "query",
          "required" : false,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "max",
          "in" : "query",
          "required" : false,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Category"
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "Category" ],
        "summary" : "Creates a category",
        "description" : "",
        "operationId" : "createCategory",
        "consumes" : [ "application/xml", "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/Category"
          }
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
    "/categories/{id}" : {
      "get" : {
        "tags" : [ "Category" ],
        "summary" : "Finds a category given an identifier",
        "description" : "",
        "operationId" : "findCategoryById",
        "produces" : [ "application/xml", "application/json" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "pattern" : "[0-9][0-9]*",
          "format" : "int64"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      },
      "put" : {
        "tags" : [ "Category" ],
        "summary" : "Updates a category",
        "description" : "",
        "operationId" : "updateCategory",
        "consumes" : [ "application/xml", "application/json" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "pattern" : "[0-9][0-9]*",
          "format" : "int64"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/Category"
          }
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      },
      "delete" : {
        "tags" : [ "Category" ],
        "summary" : "Deletes a category by id",
        "description" : "",
        "operationId" : "deleteCategoryById",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "pattern" : "[0-9][0-9]*",
          "format" : "int64"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
    "/countries" : {
      "get" : {
        "tags" : [ "Country" ],
        "summary" : "Lists all the countries",
        "description" : "",
        "operationId" : "listAllCountries",
        "produces" : [ "application/xml", "application/json" ],
        "parameters" : [ {
          "name" : "start",
          "in" : "query",
          "required" : false,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "max",
          "in" : "query",
          "required" : false,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Country"
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "Country" ],
        "summary" : "Creates a country",
        "description" : "",
        "operationId" : "createCountry",
        "consumes" : [ "application/xml", "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/Country"
          }
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
    "/countries/{id}" : {
      "get" : {
        "tags" : [ "Country" ],
        "summary" : "Retrieves a country by its id",
        "description" : "",
        "operationId" : "findCountryById",
        "produces" : [ "application/xml", "application/json" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "pattern" : "[0-9][0-9]*",
          "format" : "int64"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      },
      "put" : {
        "tags" : [ "Country" ],
        "summary" : "Updates a country",
        "description" : "",
        "operationId" : "updateCountry",
        "consumes" : [ "application/xml", "application/json" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "pattern" : "[0-9][0-9]*",
          "format" : "int64"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/Country"
          }
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      },
      "delete" : {
        "tags" : [ "Country" ],
        "summary" : "Deletes a country given an id",
        "description" : "",
        "operationId" : "deleteCountryById",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "pattern" : "[0-9][0-9]*",
          "format" : "int64"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
    "/customers" : {
      "get" : {
        "tags" : [ "Customer" ],
        "summary" : "Lists all the customers",
        "description" : "",
        "operationId" : "listAllCustomers",
        "produces" : [ "application/xml", "application/json" ],
        "parameters" : [ {
          "name" : "start",
          "in" : "query",
          "required" : false,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "max",
          "in" : "query",
          "required" : false,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Customer"
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "Customer" ],
        "summary" : "Creates a customer",
        "description" : "",
        "operationId" : "createCustomer",
        "consumes" : [ "application/xml", "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/Customer"
          }
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
    "/customers/{id}" : {
      "get" : {
        "tags" : [ "Customer" ],
        "summary" : "Finds a customer by it identifier",
        "description" : "",
        "operationId" : "findCustomerById",
        "produces" : [ "application/xml", "application/json" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "pattern" : "[0-9][0-9]*",
          "format" : "int64"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      },
      "put" : {
        "tags" : [ "Customer" ],
        "summary" : "Updates a customer",
        "description" : "",
        "operationId" : "updateCustomer",
        "consumes" : [ "application/xml", "application/json" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "pattern" : "[0-9][0-9]*",
          "format" : "int64"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/Customer"
          }
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      },
      "delete" : {
        "tags" : [ "Customer" ],
        "summary" : "Deletes a customer by id",
        "description" : "",
        "operationId" : "deleteCustomerById",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "pattern" : "[0-9][0-9]*",
          "format" : "int64"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
...
  },
  "definitions" : {
    "Address" : {
      "type" : "object",
      "required" : [ "city", "street1", "zipcode" ],
      "properties" : {
        "street1" : {
          "type" : "string",
          "minLength" : 5,
          "maxLength" : 50
        },
        "street2" : {
          "type" : "string"
        },
        "city" : {
          "type" : "string",
          "minLength" : 2,
          "maxLength" : 50
        },
        "state" : {
          "type" : "string"
        },
        "zipcode" : {
          "type" : "string",
          "minLength" : 1,
          "maxLength" : 10
        },
        "country" : {
          "$ref" : "#/definitions/Country"
        }
      }
    },
    "Category" : {
      "type" : "object",
      "required" : [ "description", "name" ],
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "version" : {
          "type" : "integer",
          "format" : "int32"
        },
        "name" : {
          "type" : "string",
          "minLength" : 1,
          "maxLength" : 30
        },
        "description" : {
          "type" : "string",
          "minLength" : 0,
          "maxLength" : 3000
        }
      }
    },
    "Country" : {
      "type" : "object",
      "required" : [ "iso3", "isoCode", "name", "numcode", "printableName" ],
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "version" : {
          "type" : "integer",
          "format" : "int32"
        },
        "isoCode" : {
          "type" : "string",
          "minLength" : 2,
          "maxLength" : 2
        },
        "name" : {
          "type" : "string",
          "minLength" : 2,
          "maxLength" : 80
        },
        "printableName" : {
          "type" : "string",
          "minLength" : 2,
          "maxLength" : 80
        },
        "iso3" : {
          "type" : "string",
          "minLength" : 3,
          "maxLength" : 3
        },
        "numcode" : {
          "type" : "string",
          "minLength" : 3,
          "maxLength" : 3
        }
      }
    },
    "Customer" : {
      "type" : "object",
      "required" : [ "firstName", "lastName", "password" ],
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "version" : {
          "type" : "integer",
          "format" : "int32"
        },
        "firstName" : {
          "type" : "string",
          "minLength" : 2,
          "maxLength" : 50
        },
        "lastName" : {
          "type" : "string",
          "minLength" : 2,
          "maxLength" : 50
        },
        "telephone" : {
          "type" : "string"
        },
        "email" : {
          "type" : "string"
        },
        "login" : {
          "type" : "string"
        },
        "password" : {
          "type" : "string",
          "minLength" : 1,
          "maxLength" : 256
        },
        "uuid" : {
          "type" : "string",
          "minLength" : 1,
          "maxLength" : 256
        },
        "role" : {
          "type" : "string",
          "enum" : [ "USER", "ADMIN" ]
        },
        "dateOfBirth" : {
          "type" : "string",
          "format" : "date"
        },
        "age" : {
          "type" : "integer",
          "format" : "int32"
        },
        "homeAddress" : {
          "$ref" : "#/definitions/Address"
        },
        "fullName" : {
          "type" : "string"
        }
      }
    },
...
  }
}